He hecho el commit para limpiar la rama

+++ getronics-spring-mvc-services/src/main/java/es/getronics/services/impl/TecnologiaServiceImpl.java
@@ -19,6 +19,7 @@ import es.getronics.dto.DepartamentoDto;
 import es.getronics.dto.EmpleadoDto;
 import es.getronics.dto.KeyValueItem;
 import es.getronics.dto.TecnologiaDto;
+import es.getronics.exceptions.ExcepcionTecnologia;
 import es.getronics.services.TecnologiaService;
 
 /**
@@ -73,19 +74,38 @@ public class TecnologiaServiceImpl implements TecnologiaService {
 	}
 
 	@Override
-	public TecnologiaDto insert(TecnologiaDto dto) {
-		if(!tecnologiaDao.findByName(dto.getNombre()).isEmpty()) {
-			throw new GetronicsDaoException("tecnologia.already.exists");
-		}
-		Tecnologia entity = tecnologiaConverter.map(dto);
-		dto = tecnologiaConverter.convert(tecnologiaDao.insert(entity));
-		
-		return dto;
-	}
+	public TecnologiaDto insert(TecnologiaDto dto) throws ExcepcionTecnologia{
+		//if(!tecnologiaDao.findByName(dto.getNombre()).isEmpty()) {
+				//	throw new GetronicsDaoException("tecnologia.already.exists");
+				//}
+				Tecnologia entity = tecnologiaConverter.map(dto);
+				
+				Boolean tecnologiaExiste=false;
+				
+				for(Tecnologia tecnologia:tecnologiaDao.findAll()) {
+					if(tecnologia.getNombre().equals(dto.getNombre())) {
+						tecnologiaExiste=true;
+					}
+				}
+				if(tecnologiaExiste) {
+					throw new ExcepcionTecnologia("La tecnologia ya existe");
+				}
+				else { 
+					dto = tecnologiaConverter.convert(tecnologiaDao.insert(entity));
+				}
+				
+				return dto;
+			}
 	
 	@Override
-	public void remove(Long id) {
-		tecnologiaDao.remove(id);
+	public void remove(Long id) throws ExcepcionTecnologia {
+		Tecnologia entity = tecnologiaDao.findById(id);
+		if (!entity.getDepartamentos().isEmpty()){
+		throw new ExcepcionTecnologia("No se puede eliminar la tecnologia porque pertenece a un Departamento");
+		}
+		else {
+			tecnologiaDao.remove(id);
+		}
 	}
 
 	@Override
